Organize map components in folder
