+++ app/assets/scripts/components/map/index.js
@@ -1,0 +1,3 @@
+import Map from './map';
+
+export default Map;
