+++ app/assets/scripts/components/map/map.js
@@ -2,9 +2,9 @@ import React, { useRef, useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import mapbox from 'mapbox-gl';
 
-import config from '../config';
+import config from '../../config';
 
-export default function Map({ style, children }) {
+export default function Map({ children }) {
   const containerRef = useRef();
 
   const [map, setMap] = useState(null);
@@ -13,15 +13,17 @@ export default function Map({ style, children }) {
     mapbox.accessToken = config.mapbox.token;
     const m = new mapbox.Map({
       container: containerRef.current,
-      style: `mapbox://styles/mapbox/light-v10/`,
+      style: config.mapbox.baseStyle,
       zoom: 1,
       center: [0, 0],
     });
 
+    m.addControl(new mapbox.NavigationControl());
+
     m.on('load', () => {
       setMap(m);
 
-      if (process.env.NODE_ENV === 'development') {
+      if (process.env.DS_ENV === 'development') {
         // makes map accessible in console for debugging
         window.map = m;
       }
@@ -35,23 +37,21 @@ export default function Map({ style, children }) {
   }, []);
 
   return (
-    <div style={style} ref={containerRef} data-cy="mapboxgl-map">
-      {map &&
-        children &&
-        React.Children.map(children, child =>
-          React.cloneElement(child, {
-            map,
-          })
-        )}
+    <div className="map">
+      <div ref={containerRef} className="map__container" data-cy="mapboxgl-map">
+        {map &&
+          children &&
+          React.Children.map(children, child =>
+            React.cloneElement(child, {
+              map,
+            })
+          )}
+      </div>
     </div>
   );
 }
 
 Map.propTypes = {
-  style: PropTypes.shape({
-    height: PropTypes.oneOfType([PropTypes.number, PropTypes.string])
-      .isRequired,
-  }),
   children: PropTypes.oneOfType([
     PropTypes.element,
     PropTypes.arrayOf(PropTypes.element),
